+++ src/components/index.js
@@ -1,0 +1,4 @@
+export {default as TitleSection} from './TitleSection';
+export {default as Icon} from './Icon';
+
+export {default as AppHeader} from './Header';
