+++ src/components/tables/index.js
@@ -1,0 +1,3 @@
+export {default as ServiceTable} from './ServiceTable';
+export {default as CommitTable} from './CommitTable';
+export {default as DeployTable} from './DeployTable';
