+++ src/pages/index.js
@@ -1,0 +1,8 @@
+export {default as Home} from './Home';
+export {default as Services} from './Services';
+export {default as Commits} from './Commits';
+export {default as Deploys} from './Deploys';
+
+export {default as Service} from './Service';
+
+export {default as Error} from './Error';
